Basic setup for transaction page
This commit only works if the backend considers the values supplied
for the transaction's person and individual-product as simply strings
(e.g. in mu-cl-resources: `:properties` of the transaction), not having
any relationship with resources person and individual-product yet.

Later commits will add relationships to the frontend to make it work
with a backend that has relationships defined.